refactor: move geometry components to example folder
