+++ examples/geometry/src/components/index.ts
