+++ src/index.ts
@@ -5,7 +5,6 @@ export * from './components/map';
 export * from './components/map-control';
 export * from './components/marker';
 export * from './components/pin';
-export * from './components/geometry';
 export * from './hooks/use-api-loading-status';
 export * from './hooks/use-api-is-loaded';
 export * from './hooks/use-maps-library';
@@ -15,3 +14,4 @@ export * from './hooks/directions-service';
 export * from './hooks/street-view-panorama';
 
 export {limitTiltRange} from './libraries/limit-tilt-range';
+export * from './libraries/lat-lng-utils';
