--- src/libraries/is-lat-lng-literal.ts
@@ -1,8 +1,0 @@
-export function isLatLngLiteral(
-  obj: unknown
-): obj is google.maps.LatLngLiteral {
-  if (!obj || typeof obj !== 'object') return false;
-  if (!('lat' in obj && 'lng' in obj)) return false;
-
-  return Number.isFinite(obj.lat) && Number.isFinite(obj.lng);
-}
